refactor(BackUpLoader): Splits the functionality in "raw" and "compress" subcommands
Now "raw" will handle the uncompressed uploads of folder and files while "compress" will create the .tar.gz archive and upload the latter to the remote location.
Both the subcommands will have a --username and --host argument to avoid prompt insertion.
If not provided then the programs will ask for both and the password via prompt.