build(gtts): Adds gtts as dependecy for future usage
This commits adds gtts (version 2.2.4) as dependency for future usage.

The future script will try to read a given pdf file, extracts the text from it and convert it via Google' Text To Speech API to an audio (mp3) stream that will be either saved locally via "export" subcommand or streamed/played to the user via the "stream" subcommand.