feat(DocReader): Initial implementation of DocReader
This implementation is very unstable and still experimental.
As said before the goal is to read the content from a PDF file and convert it to an audio stream.
At the moment much of the conversion part is done  and completely handled by Google's Text To Speech API, now the only two missing feature is the text extraction from the PDF and the "live" reproduction of the stream, once converted, to the user (via the "stream" subcommand).