fix(DocReader): Fixed a bug that caused the audio output to be full of gibberish
Since texttract returns a buffer of bytes (binary format), while I thoght it was a string, some strange artifacts presented in the final audio. (strange, non natural words), this was caused by the fact that to gTTS was given a binary format while expecting a str, and without any context the buffer was casted/transformed to the best of gTTS options.

For example the character "\n" were interpreted not as a whitespace but as two character "\" and "n" and given to the gTTS API that translated up to its best.

Now the output from textract is converted to a str with the correct enconding and the gTTS output presents much less gibberish.