fix(DocReader): Implemented the "stream" subcommand
While I would've loved to do a more real time approach with the script writing on a buffer and another agent reading from it concurrently, at the moment it seems quite advanced and requires to install some heavy dependencies such as PyAudio as well as implementing the whole PLayer TUI by myself.
To avoid this I switched to  a more simplistic approach in which the audio is saved in temporary file (similarly to "export"), once this first step is completed nlvc (the terminal version of VLC) is started with the same file as input,  starting the reproduction.

In future it will be possible to select other tools/commands instead of being forced to nvlc, this will allow for more customization as well as avoidcrashing when nvlc is not installed.