fix(test_youtubescraper): Fixed some errors and bugs in test suite
With a proper working version of PyTube I was able to test my test suite and correct some minor errors in the code.
I also removed caption related test since the functionality doesn't work correctly in pytube as of now, once fixed and available the test case will be reintroduced.