feat(DocReader): Adds player_flag forwarding to "stream" subcommand
Now it is possible, when using the "stream" subcommand in DocReader.py, to define one or more flags that will be
forwarded to the audio player program of choice before execution.
An example with nvlc will be that with the --play-and exit flag forwarded I can start a player that will automatically close
upon completion.