fix(test): Changed pytest configuration to condider unit_ prefix
Since pytest, when executing without specified input files, searches by default all the files with the "test_" prefix
it is not possible anymore to use "poetry run pytest" to bulk execute all the test suites.
This commit fixes the misconfiguration by manually adding to pyproject.toml a config override and directly specifying to
pytest to consider all "unit_*" file to be test files.

Adds as well poetry.lock that somehow was missing