ci(init): First draft of CI action
This first draft of the GitHub Action should implement a "CI Testing" that, as the name suggest should run the whole
test suite when a pull request is made to the project.
If one of the test cases fails then the whole workflow should fail and report an error on the PR details.