ci(env): Added env variable mapped to GitHub's secrets
Some test suites, namely test_backuploader.py, need to access via SSH
a remote location.

When running the tests locally the access credentials are retrieved thorugh
some environment variables that the user has to setup accordingly, else the
test suite fails.

In the case of the CI pipeline the "private" environment variables (such as
hostname and password) are mapped to GitHub secret so that the test suite
can be executed in  full but without leaking any private information.