feat(DriveDiffMerger): Made further improvements
The current strategy to pull/update from Google Drive seems to have some shortcomings.
As of now, it's quite difficult to manage only the pulling from Drive, an eventual
push to Drive would be madness.

I'm thinking  about changing the strategy completely, first traversing the local and remote
tree simultaneously, extracting a diff that will keep track of the changes to be applied.
Said diff will be applied in both directions by an executor.