ci(Ci Testing): Fixed 'missing GitHub secrets' error
Before CI pipeline failed because the GitHub secret didn't get "expanded".

This was related to the fact that pipelines which runs on "pull_request"
events are executed in untrusted mode due to the risk of malicious code push.
Given that now CI Testing will be executed only on push to main (after the
merge of the PR) and the PR assignee will have to eventually execute the test locally.

Reference: https://github.com/pypa/gh-action-pypi-publish/discussions/49